renamed shader chunks
