+++ src/scene/shader-lib/glsl/chunks/gsplat/frag/gsplatSogsReorder.js
