+++ src/scene/shader-lib/wgsl/chunks/gsplat/frag/gsplatSogsReorder.js
